requirements.txt
flask==2.3.3
google-auth-oauthlib==1.0.0
google-auth-httplib2==0.1.0
google-api-python-client==2.97.0
sentence-transformers==2.2.2
faiss-cpu==1.7.4
python-dotenv==1.0.0
transformers==4.31.0
torch==2.0.1
pandas==2.0.3 